Refactor: date picker code seperation
