+++ src/features/booth/index.ts
@@ -7,3 +7,15 @@ export { DeleteButton as BoothDeleteButton } from "./ui/DeleteButton";
 export { EditButton as BoothEditButton } from "./ui/EditButton";
 export { BoothAvailabilitySwitchButton as BoothSwitchButton } from "./ui/BoothAvailabilitySwitchButton";
 export { Button as AddBoothButton } from "./ui/Button";
+
+// UI components
+export { OperatingDaysCalendar } from "./ui/operating-time/OperatingDaysCalendar";
+export { OperatingTimeInput } from "./ui/operating-time/OperatingTimeInput";
+
+// Model/Types
+export type { OperatingTime } from "./model/operating-time";
+export {
+  convertToHHMM,
+  convertToHHMMSS,
+  formatDateDisplay,
+} from "./model/operating-time";
